+++ python/geoarrow-core/src/ffi/to_python/array.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use geoarrow::array::GeometryArrayDyn;
+use geoarrow::array::NativeArrayDyn;
 use geoarrow::chunked_array::ChunkedGeometryArrayTrait;
 use geoarrow::NativeArray;
 use pyo3::prelude::*;
@@ -12,7 +12,7 @@ pub fn geometry_array_to_pyobject(
     py: Python,
     arr: Arc<dyn NativeArray>,
 ) -> PyGeoArrowResult<PyObject> {
-    Ok(PyGeometryArray::new(GeometryArrayDyn::new(arr)).into_py(py))
+    Ok(PyGeometryArray::new(NativeArrayDyn::new(arr)).into_py(py))
 }
 
 pub fn chunked_geometry_array_to_pyobject(