+++ python/pyo3-geoarrow/src/array.rs
@@ -4,7 +4,7 @@ use crate::error::{PyGeoArrowError, PyGeoArrowResult};
 use crate::{PyGeometry, PyGeometryType};
 use arrow::datatypes::Schema;
 use arrow_array::RecordBatch;
-use geoarrow::array::{from_arrow_array, GeometryArrayDyn};
+use geoarrow::array::{from_arrow_array, NativeArrayDyn};
 
 use geoarrow::error::GeoArrowError;
 use geoarrow::scalar::GeometryScalar;
@@ -19,10 +19,10 @@ use pyo3_arrow::ffi::to_array_pycapsules;
 use pyo3_arrow::PyArray;
 
 #[pyclass(module = "geoarrow.rust.core._rust", name = "GeometryArray", subclass)]
-pub struct PyGeometryArray(pub(crate) GeometryArrayDyn);
+pub struct PyGeometryArray(pub(crate) NativeArrayDyn);
 
 impl PyGeometryArray {
-    pub fn new(array: GeometryArrayDyn) -> Self {
+    pub fn new(array: NativeArrayDyn) -> Self {
         Self(array)
     }
 
@@ -39,7 +39,7 @@ impl PyGeometryArray {
         self.0.as_ref()
     }
 
-    pub fn into_inner(self) -> GeometryArrayDyn {
+    pub fn into_inner(self) -> NativeArrayDyn {
         self.0
     }
 
@@ -147,19 +147,19 @@ impl PyGeometryArray {
     }
 }
 
-impl From<GeometryArrayDyn> for PyGeometryArray {
-    fn from(value: GeometryArrayDyn) -> Self {
+impl From<NativeArrayDyn> for PyGeometryArray {
+    fn from(value: NativeArrayDyn) -> Self {
         Self(value)
     }
 }
 
 impl From<NativeArrayRef> for PyGeometryArray {
     fn from(value: NativeArrayRef) -> Self {
-        Self(GeometryArrayDyn::new(value))
+        Self(NativeArrayDyn::new(value))
     }
 }
 
-impl From<PyGeometryArray> for GeometryArrayDyn {
+impl From<PyGeometryArray> for NativeArrayDyn {
     fn from(value: PyGeometryArray) -> Self {
         value.0
     }